+++ src/simAutoReg-fit-var-ols.cpp
@@ -13,46 +13,12 @@
 //' @param Y Numeric matrix.
 //'   Matrix of dependent variables (Y).
 //' @param X Numeric matrix.
-//'   Matrix of lagged predictors (X).
+//'   Matrix of predictors (X).
 //'
 //' @return Matrix of estimated autoregressive and cross-regression coefficients.
 //'
 //' @examples
-//' set.seed(42)
-//' time <- 100000L
-//' burn_in <- 200
-//' k <- 3
-//' p <- 2
-//' constant <- c(1, 1, 1)
-//' coef <- matrix(
-//'   data = c(
-//'     0.4, 0.0, 0.0, 0.1, 0.0, 0.0,
-//'     0.0, 0.5, 0.0, 0.0, 0.2, 0.0,
-//'     0.0, 0.0, 0.6, 0.0, 0.0, 0.3
-//'   ),
-//'   nrow = k,
-//'   byrow = TRUE
-//' )
-//' chol_cov <- chol(
-//'   matrix(
-//'     data = c(
-//'       0.1, 0.0, 0.0,
-//'       0.0, 0.1, 0.0,
-//'       0.0, 0.0, 0.1
-//'     ),
-//'     nrow = k,
-//'     byrow = TRUE
-//'   )
-//' )
-//' y <- SimVAR(
-//'   time = time,
-//'   burn_in = burn_in,
-//'   constant = constant,
-//'   coef = coef,
-//'   chol_cov = chol_cov
-//' )
-//' yx <- YX(y, p)
-//' FitVAROLS(Y = yx$Y, X = yx$X)
+//' FitVAROLS(Y = VAR_YX$Y, X = VAR_YX$X)
 //'
 //' @details
 //' The [simAutoReg::FitVAROLS()] function estimates the parameters
@@ -60,14 +26,16 @@
 //' using the Ordinary Least Squares (OLS) method.
 //' Given the input matrices `Y` and `X`,
 //' where `Y` is the matrix of dependent variables,
-//' and `X` is the matrix of lagged predictors,
+//' and `X` is the matrix of predictors,
 //' the function computes the autoregressive and cross-regression coefficients
 //' of the VAR model.
+//' Note that if the first column of `X` is a vector of ones,
+//' the constant vector is also estimated.
 //'
 //' The steps involved in estimating the VAR model parameters using OLS are as follows:
 //'
 //' - Compute the QR decomposition of the lagged predictor matrix `X`
-//'   using the `qr_econ` function from the Armadillo library.
+//'   using the `qr` function from the Armadillo library.
 //' - Extract the `Q` and `R` matrices from the QR decomposition.
 //' - Solve the linear system `R * coef = Q.t() * Y`
 //'   to estimate the VAR model coefficients `coef`.
@@ -75,56 +43,20 @@
 //'   autoregressive and cross-regression coefficients of the VAR model.
 //'
 //' @seealso
-//' The `qr_econ` function from the Armadillo library for QR decomposition.
+//' The `qr` function from the Armadillo library for QR decomposition.
 //'
 //' @importFrom Rcpp sourceCpp
 //'
 //' @export
 // [[Rcpp::export]]
-arma::mat FitVAROLS(arma::mat Y, arma::mat X) {
+arma::mat FitVAROLS(const arma::mat& Y, const arma::mat& X)
+{
   // Estimate VAR model parameters using QR decomposition
   arma::mat Q, R;
-  arma::qr_econ(Q, R, X);
+  arma::qr(Q, R, X);
 
   // Solve the linear system R * coef = Q.t() * Y
   arma::mat coef = arma::solve(R, Q.t() * Y);
 
   return coef.t();
 }
-
-/*** R
-time <- 100000L
-burn_in <- 200
-k <- 3
-p <- 2
-constant <- c(1, 1, 1)
-coef <- matrix(
-  data = c(
-    0.4, 0.0, 0.0, 0.1, 0.0, 0.0,
-    0.0, 0.5, 0.0, 0.0, 0.2, 0.0,
-    0.0, 0.0, 0.6, 0.0, 0.0, 0.3
-  ),
-  nrow = k,
-  byrow = TRUE
-)
-chol_cov <- chol(
-  matrix(
-    data = c(
-      0.1, 0.0, 0.0,
-      0.0, 0.1, 0.0,
-      0.0, 0.0, 0.1
-    ),
-    nrow = k,
-    byrow = TRUE
-  )
-)
-y <- SimVAR(
-  time = time,
-  burn_in = burn_in,
-  constant = constant,
-  coef = coef,
-  chol_cov = chol_cov
-)
-yx <- YX(y, p)
-FitVAROLS(Y = yx$Y, X = yx$X)
-*/